refactor(transit): extract stop snapshot helper
